% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomadsNC.R
\docType{package}
\name{nomadsNC}
\alias{nomadsNC}
\alias{nomadsNC-package}
\title{nomadsNC}
\description{
\code{nomadsNC} package
}
\details{
Build usable flow matrices or netcdf files from NOMADS

See the README on [github](https://github.com/mikejohnson51/nomadsNC)
}
